Tried to add is_done checkbox, but only shows in client

+++ todo/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import redirect, render
 from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from todo.models import Todo
 
 
@@ -14,9 +15,11 @@ def index(request):
         return HttpResponse('Invalid request method', status=405)
 
 
+@login_required(login_url='/user/login/')
 def create(request):
     if request.method == 'POST':
-        Todo.objects.create(content=request.POST['content'], user=request.user)
+        Todo.objects.create(
+            content=request.POST['content'], user=request.user)
         return redirect('/todo/')
     elif request.method == 'GET':
         return render(request, 'todo/create.html')
@@ -34,9 +37,12 @@ def read(request, todo_id):
 
 def delete(request, todo_id):
     if request.method == 'POST':
-        todo = Todo.objects.get(id=todo_id)
-        todo.delete()
-        return redirect('/todo/')
+        if request.user == todo.user:
+            todo = Todo.objects.get(id=todo_id)
+            todo.delete()
+            return redirect('/todo/')
+        else:
+            return HttpResponse('You are not allowed to delete this todo', status=403)
     else:
         return HttpResponse('Invalid request method', status=405)
 
@@ -49,9 +55,23 @@ def update(request, todo_id):
         }
         return render(request, 'todo/update.html', context)
     elif request.method == 'POST':
+        if request.user == todo.user:
+            todo = Todo.objects.get(id=todo_id)
+            todo.content = request.POST['content']
+            todo.save()
+            return redirect(f'/todo/{todo_id}/')
+        else:
+            return HttpResponse('You are not allowed to update this todo', status=403)
+    else:
+        return HttpResponse('Invalid request method', status=405)
+
+
+def isdone(request, todo_id):
+    if request.method == 'POST':
         todo = Todo.objects.get(id=todo_id)
-        todo.content = request.POST['content']
-        todo.save()
-        return redirect(f'/todo/{todo_id}/')
+        todo.is_done = request.POST['is_done']
+        return redirect('/todo/')
+    elif request.method == 'GET':
+        return render(request, 'todo/create.html')
     else:
         return HttpResponse('Invalid request method', status=405)
